Encapsulate link click handler code in new component and convert page that use legacy webfield components to use it
